optimized and updated the contract sol file
Hey samjohn, great Dapp. I went through the codebase and noticed no bud. However as Brindrajsinh suggested and optimized the contract by removing the _isHired and creating a constructor to handle the owner and fee. I kept all changes and add documentations. Next i implementated/added the chat platform for the freelancer and client to communicate.