optimized and updated the contract sol file
Hey samjohn, great Dapp. I went through the codebase and noticed no bud. However as Brindrajsinh suggested and optimized the contract by removing the _isHired and creating a constructor to handle the owner and fee. I kept all changes and add documentations. Next i implementated/added the chat platform for the freelancer and client to communicate. To do that i first created addresses and messageAmounts mapping and a new Message struct. Created a getReceivedMessageCount function which gets the length of the received message mapping of an adress and a getSentMessageCount function which gets the length of the sent message mapping of an adress. I created a getReceivedMessage function which gets a received message based on index and a lot more functions like the uint2str, getSentMessage e.t.c. Please check the change on the PR below. Thanks